splitting resources provider to its own file
